Move KDF Vectors testing to their own file
